+++ src/main/java/com/project/shopping/config/WebMvcConfig.java
@@ -1,6 +1,8 @@
 package com.project.shopping.config;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.cors.CorsUtils;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -8,6 +10,12 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
     private  final long MAX_AGE_SECS = 3600;
+
+    // WebSecurityConfig 에서 사용시 WebSecurityConfig -> CustomOAuth2UserService -> UserService 에서 순환 참조 현상 발생
+    @Bean
+    public BCryptPasswordEncoder passwordEncoder(){
+        return new BCryptPasswordEncoder();
+    }
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
